chore: install prisma/client-runtime-utils (#523)
* chore: install prisma/client-runtime-utils

we have a runtime lint/typecheck issue where tsc can't resolve the type
properly.

refer to this issue:
https://github.com/prisma/prisma/issues/28581#issuecomment-3575082428,
where the problem exists due to symlinks used by pnpm to link to
packages. to avoid this, we have to install the required prisma package
at the package's top level

* chore: copy env prior to typecheck

prisma now uses the prisma config file, which calls into `env` to get
teh database url to generate types.

this requires the env var to be set to work, so we copy the example env
to the env

* chore: patch prisma/client

we need to patch the client so that the missing types are available